common callnum methods isolated
